requirements.txt
alembic==1.4.2
click==7.1.2
dominate==2.5.2
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.17.0
Mako==1.1.3
MarkupSafe==1.1.1
numpy==1.19.4
psycopg2-binary==2.8.6
python-dateutil==2.8.1
python-editor==1.0.4
scikit-learn==0.23.2
scipy==1.5.4
six==1.15.0
sklearn==0.0
SQLAlchemy==1.3.19
threadpoolctl==2.1.0
visitor==0.1.3
Werkzeug==1.0.1
WTForms==2.3.3
